[수정] 파일구조 개선 (2023.03.10 이상윤)
