+++ udog_server/src/main/java/com/mainproject/udog_server/api/dog/repository/DogRepository.java
@@ -1,0 +1,7 @@
+package com.mainproject.udog_server.api.dog.repository;
+
+import com.mainproject.udog_server.api.dog.entity.Dog;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface DogRepository extends JpaRepository<Dog, Long> {
+}
