+++ udog_server/src/main/java/com/mainproject/udog_server/api/reviewLike/controller/ReviewLikeController.java
@@ -1,0 +1,4 @@
+package com.mainproject.udog_server.api.reviewLike.controller;
+
+public class ReviewLikeController {
+}
