+++ udog_server/src/main/java/com/mainproject/udog_server/api/reviewLike/mapper/ReviewLikeMapper.java
@@ -1,0 +1,4 @@
+package com.mainproject.udog_server.api.reviewLike.mapper;
+
+public interface ReviewLikeMapper {
+}
