+++ udog_server/src/main/java/com/mainproject/udog_server/api/reviewLike/repository/ReviewLikeRepository.java
@@ -1,0 +1,4 @@
+package com.mainproject.udog_server.api.reviewLike.repository;
+
+public interface ReviewLikeRepository {
+}
