+++ udog_server/src/main/java/com/mainproject/udog_server/api/reviewLike/service/ReviewLikeService.java
@@ -1,0 +1,4 @@
+package com.mainproject.udog_server.api.reviewLike.service;
+
+public class ReviewLikeService {
+}
