--- udog_server/src/main/java/com/mainproject/udog_server/dog/entity/Dog.java
@@ -1,45 +1,0 @@
-package com.mainproject.udog_server.dog.entity;
-
-import lombok.*;
-
-import javax.persistence.*;
-import java.time.LocalDate;
-
-@Entity
-@Getter
-@Setter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Dog {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long dogId;
-
-    @Column(nullable = false, length = 100)
-    private String dogName;
-
-    @Column(nullable = false, length = 20)
-    private LocalDate dogBirthDate;
-
-    @Column(nullable = false, length = 25)
-    private String dogSpecies;
-
-    @Column(nullable = false, length = 20)
-    private int dogWeight;
-
-    @Column
-    private String dogDescription;
-
-//    @ManyToOne(fetch = FetchType.LAZY)
-//    @JoinColumn(name = "member_id")
-//    private Member member;
-
-    @Builder
-    public Dog (long dogId, String dogName, LocalDate dogBirthDate, String dogSpecies, int dogWeight, String dogDescription) {
-        this.dogId = dogId;
-        this.dogName = dogName;
-        this.dogBirthDate = dogBirthDate;
-        this.dogSpecies = dogSpecies;
-        this.dogWeight = dogWeight;
-        this.dogDescription = dogDescription;
-    }
-}
