--- udog_server/src/main/java/com/mainproject/udog_server/reviewLike/entity/ReviewLike.java
@@ -1,4 +1,0 @@
-package com.mainproject.udog_server.reviewLike.entity;
-
-public class ReviewLike {
-}
