--- udog_server/src/main/java/com/mainproject/udog_server/reviewLike/mapper/ReviewLikeMapper.java
@@ -1,4 +1,0 @@
-package com.mainproject.udog_server.reviewLike.mapper;
-
-public interface ReviewLikeMapper {
-}
