--- udog_server/src/main/java/com/mainproject/udog_server/reviewLike/repository/ReviewLikeRepository.java
@@ -1,4 +1,0 @@
-package com.mainproject.udog_server.reviewLike.repository;
-
-public interface ReviewLikeRepository {
-}
