--- udog_server/src/main/java/com/mainproject/udog_server/reviewLike/service/ReviewLikeService.java
@@ -1,4 +1,0 @@
-package com.mainproject.udog_server.reviewLike.service;
-
-public class ReviewLikeService {
-}
